Upload function now running

+++ database/migrations/2021_05_08_125924_add_column_to_user.php
@@ -12,7 +12,7 @@ class AddColumnToUser extends Migration {
      */
     public function up() {
         Schema::table('users', function (Blueprint $table) {
-            $table->integer('role');
+            $table->tinyInteger('role');
             $table->string('lastname')->nullable();
             $table->string('patronymic')->nullable();
             $table->integer('organ')->nullable();
